% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_DiffR2.R
\name{DCP_DiffR2}
\alias{DCP_DiffR2}
\title{Differential Rhythmicity Fitness Test}
\usage{
DCP_DiffR2(
  x,
  method = "LR",
  TOJR = NULL,
  alpha = 0.05,
  nSampling = 1000,
  Sampling.save = NULL,
  Sampling.file.label = "gII_vs_gI",
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{outpuf of DCP_Rhythmicity(x1, x2)}

\item{method}{one of "LR", "permutation", "bootstrap". "LR" is recommended.}

\item{TOJR}{toTOJR output. If NULL, rhythm.joint object in x will be used.}

\item{alpha}{cutoff of p-values for significant differential rhythm fitness change.}

\item{nSampling}{number of samplings if "permutation" or "bootstrap" is chosen for method}

\item{Sampling.save}{directory to save sampling results. If NULL, the result will not be saved.}

\item{Sampling.file.label}{character string. Used as file lable if sampling results are saved.}

\item{p.adjust.method}{input for p.adjust() in R package \code{stat}}

\item{parallel.ncores}{integer. Number of cores used if using parallel computing with \code{mclapply()}. Not functional for windows system.}
}
\value{
A dataframe of differential rhythm finess test results.
}
\description{
Differential Rhythmicity Fitness Test
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
rhythm.res = DCP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.diffR2 = DCP_DiffPar(rhythm.res)
}
