% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_getZT.R
\name{DCP_getZT}
\alias{DCP_getZT}
\title{Title Calculate Zeitgeber Time}
\usage{
DCP_getZT(t, lat, long, ZT.min = -6)
}
\arguments{
\item{t}{Recorded time in POSIXlt format. Please make sure that the input time has the correct time zone.}

\item{lat}{Local latitude corresponding to the recorded time.}

\item{long}{Local longitude corresponding to the recorded time.}

\item{ZT.min}{The mininum of output ZT. The default is -6, which corresponds to output range of (-6, 18).}
}
\value{
Zeitgeiber time corrected by local sunrise
}
\description{
Calculate Zeitgeber Time based on clock time based on local sunrise time.
}
\examples{
t = as.POSIXlt("2022-09-17 13:43:00 EST", tz="America/New_York",usetz=TRUE)
lat = 40.4406; long = -79.9959 #This is the Pittsburgh coordinates
t.correct = DCP_getZT(t, lat, long, ZT.min = -6)
}
