% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_GraphFunctions.R
\name{DCP_PlotHeatmap}
\alias{DCP_PlotHeatmap}
\title{Heatmaps of Rhythmicity Signals}
\usage{
DCP_PlotHeatmap(
  x,
  genes.plot = NULL,
  col_breaks = c(seq(-2, 0, length = 100), seq(0.1, 1, length = 200), seq(1.1, 2,
    length = 200)),
  col_low_mid_high = c("blue", "yellow", "gold"),
  Info1 = "group I",
  Info2 = "group II",
  ...,
  filename = NULL,
  file.width = 8,
  file.height = 4
)
}
\arguments{
\item{x}{DCP_Rhythmicity() output}

\item{genes.plot}{a vector of character strings. Names of genes to be plotted. Names should be same in nomenclature as gname in the data list. If NULL, the top 100 most rhythmic genes from group I will be used}

\item{col_breaks}{color breaks}

\item{col_low_mid_high}{color choice for the heatmaps}

\item{Info1}{character string. Used in the plot title for group I.}

\item{Info2}{character string. Used in the plot title for group II (if exist).}

\item{...}{other argument to pass to \code{ComplexHeatmap::Heatmap()}}

\item{filename}{character string of the filename for exporting. If NULL, plots are not saved.}

\item{file.width}{height of the export plot}

\item{file.height}{width of the export plot}
}
\description{
Heatmaps of Rhythmicity Signals
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)

rhythm.res = DCP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
DCP_PlotHeatmap(rhythm.res)
}
