% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_GraphFunctions.R
\name{DCP_PlotPeakLink}
\alias{DCP_PlotPeakLink}
\title{Circos plot for linked peak time
Make a circos plot for linked peak time between two groups}
\usage{
DCP_PlotPeakLink(
  x,
  TOJR,
  dPhase,
  color.cut = list(param = "pvalue", fun = "<", val = 0.05, color.sig = "#03a1fc",
    color.none = "dark grey"),
  color.df = NULL,
  time.start = -6,
  Info1 = "groupI",
  Info2 = "groupII",
  filename = NULL,
  file.width = 8,
  file.height = 8,
  cir.y.breaks = seq(-6, 18, 4),
  axis.text.size = 12,
  legend.position = "right",
  color.link.low = "#e8f7ff",
  color.link.high = "#03a1fc"
)
}
\arguments{
\item{x}{DCP_Rhythmicity() output.}

\item{TOJR}{toTOJR output. If NULL, rhythm.joint object in x will be used.}

\item{dPhase}{DCP_DiffPar() output with phase shift tested.}

\item{color.cut}{list. Genes will be plotted according a cutoff. Used only when dPhase is not NULL. \itemize{
\item param parameter used for the cutoff. Should be a column in dPhase.
\item fun character string. Either "<", or ">"
\item val numeric. The value used for the cutoff
\item color.sig color for points that pass the color.cut criterion.
\item color.none color for points that do not pass the color.cut criterion.}}

\item{color.df}{data.frame. A more advanced color coding. The data frame should have two columns named color and label. label will be displayed as legend. The color vector should have the same order as genes in dPhase.}

\item{time.start}{numeric. What time do you want the phase start? Default is -6, which is midnight if time is in ZT scale.}

\item{Info1}{character string. Used in the plot title for group I}

\item{Info2}{character string. Used in the plot title for group II (if exist).}

\item{filename}{character string. The filename for plot export. If NULL, the plot will not be saved.}

\item{file.width}{width of the export plot}

\item{file.height}{height of the export plot}

\item{cir.y.breaks}{numeric. A vector for breaks for the angles. Should start with time.start and end with time.start+period}

\item{axis.text.size}{numeric. Size for the axis text.}

\item{legend.position}{One of "left”, "top", "right", "bottom", or "none"}

\item{color.link.low}{color of the link for genes with smallest |peak difference|}

\item{color.link.high}{color of the link for genes with highest |peak difference|}
}
\value{

}
\description{
Circos plot for linked peak time
Make a circos plot for linked peak time between two groups
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(2, 3), A2=c(2, 3),
phase1=c(0, pi/4), phase2=c(pi/2, pi*3/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
rhythm.res = DCP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.diffPar = DCP_DiffPar(rhythm.res, "phase")
DCP_PlotPeakLink(rhythm.res, NULL, rhythm.diffPar)

}
