% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_sim_data.R
\name{DCP_sim_data}
\alias{DCP_sim_data}
\title{Simulating Data for Rhythmicity Comparison}
\usage{
DCP_sim_data(
  ngene = 1000,
  nsample = 30,
  A1 = c(1, 3),
  A2 = c(1, 3),
  phase1 = c(0, 2 * pi),
  phase2 = c(0, 2 * pi),
  M1 = c(4, 6),
  M2 = c(4, 6),
  sigma1 = 1,
  sigma2 = 1
)
}
\arguments{
\item{ngene}{integer. Number of features. Shared by both groups}

\item{nsample}{integer. Number of samples Shared by both groups}

\item{A1}{vector of length 2. c(min, max) amplitude for group I}

\item{A2}{vector of length 2. c(min, max) amplitude for group II}

\item{phase1}{vector of length 2. c(min, max) phase for group I (values not in the 0 to \eqn{2\pi} interval will be converted)}

\item{phase2}{vector of length 2. c(min, max) phase for group II}

\item{M1}{vector of length 2. c(min, max) MESOR for group I}

\item{M2}{vector of length 2. c(min, max) MESOR for group I}

\item{sigma1}{numeric. Noise level for group I.}

\item{sigma2}{numeric. Noise level for group II.}
}
\value{
A list of length two, each is data for one group.
}
\description{
\code{DCP_sim_data} simulates a simple two-group study. The simulated data is just used for illustration purposes and is not intended to mimic real data.
}
\examples{
#simulate data with shifted phase
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
}
