% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_summaryFunctions.R
\name{SummarizeDR}
\alias{SummarizeDR}
\title{Summary of Differential Rhythmicity Tests
Summarize differential rhythmicity Tests results into a table.}
\usage{
SummarizeDR(result, test = "DRF", type = "p-value", val = 0.05, out = "long")
}
\arguments{
\item{result}{a list of vectors, usually p-values. Each vector will be summarized with a different cutoff.}

\item{test}{a vector of strings indicating the name of tests.}

\item{type}{a vector of strings indicating the type of the results.}

\item{val}{The cutoff used to dichotomize the result vectors.}

\item{out}{Should the output table be "long" or "wide"?}
}
\value{
a table of summarized results
}
\description{
Summary of Differential Rhythmicity Tests
Summarize differential rhythmicity Tests results into a table.
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
rhythm.res = DCP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.diffR2 = DCP_DiffPar(rhythm.res)
SummarizeDR(result = list(rhythm.diffR2$p.R2, rhythm.diffR2$p.R2, rhythm.diffR2$q.R2),
test = c(rep("DRF", 2), "DRF"),
type = c(rep("p-value", 2), "q-value"),
val = c(0.1, 0.05, 0.4), "long")
}
