% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCP_GraphFunctions.R
\name{DCP_PlotDisplay}
\alias{DCP_PlotDisplay}
\title{Displaying DCP_ScatterPlot outputs}
\usage{
DCP_PlotDisplay(
  x = DCP_ScatterPlot(x, genes.plot = NULL, Info1 = "gI", Info2 = "gII", filename =
    NULL, height = 8, width = 8),
  id = 1:2
)
}
\arguments{
\item{x}{DCP_ScatterPlot output}

\item{id}{integer. Indexes for plots to display.}
}
\description{
Displaying DCP_ScatterPlot outputs
}
